pyproject.toml
[project]
name = "mysql-transfer"
version = "0.1.0"
description = "CLI tool to transfer MySQL databases between source and destination servers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pymysql>=1.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "sshtunnel>=0.4",
    "inquirerpy>=0.3.4",
    "paramiko<4.0",
]

[project.scripts]
mysql-transfer = "mysql_transfer.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mysql_transfer"]
